CarlosMassa/requirements.txt
requests>=2.13.0
python-firebase==1.2
pymongo==3.4.0